move reliance edge functions to seperate file
