+++ Project/FileSystem/FileSystemTasks.c
@@ -15,6 +15,75 @@
 #include <redvolume.h>
 #include <redtests.h>
 
+//File System includes
+/*
+#include <redfs.h>
+#include <redposix.h>
+#include <redfse.h>
+#include <redconf.h>
+#include <redvolume.h>
+#include <redtests.h>
+
+//Custom FileSystem tasks
+#include "FileSystemTasks.h"
+*/
+
+
+// Arguments to mount redfs filesystems
+struct redfs_args
+{
+	const char *fspec;  // Name of device to mount
+	unsigned int flags; // REDFSMNT_* flags
+};
+
+#define REDFSMNT_FMTONCE 0x1U   // format if volume cannot be mounted
+#define REDFSMNT_FMTALWAYS 0x2U // format always
+
+
+//example how to start File system
+/*
+	int32_t iErr;
+	const char *pszVolume0 = gaRedVolConf[0].pszPathPrefix;
+	const char *pszVolume1 = gaRedVolConf[1].pszPathPrefix;
+
+	iErr = red_init();
+	if (iErr == -1)
+	{
+		fprintf(stderr, "Unexpected error %d from red_init()\n", (int)red_errno);
+		exit(red_errno);
+	}
+
+	iErr = red_format(pszVolume0);
+	if (iErr == -1)
+	{
+		fprintf(stderr, "Unexpected error %d from red_format()\n", (int)red_errno);
+		exit(red_errno);
+	}
+
+	iErr = red_mount(pszVolume0);
+	if (iErr == -1)
+	{
+		fprintf(stderr, "Unexpected error %d from red_mount()\n", (int)red_errno);
+		exit(red_errno);
+	}
+
+	iErr = red_format(pszVolume1);
+	if (iErr == -1)
+	{
+		fprintf(stderr, "Unexpected error %d from red_format()\n", (int)red_errno);
+		exit(red_errno);
+	}
+
+	iErr = red_mount(pszVolume1);
+	if (iErr == -1)
+	{
+		fprintf(stderr, "Unexpected error %d from red_mount()\n", (int)red_errno);
+		exit(red_errno);
+	}
+
+	//FileSystem Tasks
+	vStartFSWriteTask(mainCREATOR_TASK_PRIORITY);
+*/
 void foo(void *p)
 {
 	//remove compiler warning about unused parameters
